pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "wiki"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "PyYAML",
    "requests",
]

[project.scripts]
wiki = "wiki.cli:main"

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
wiki = ["prompts/*.md"]
